Channels 2 support
These changes (with the ones I propose for server.py) give beatserver Channels 2 support without dropping Channels 1 support. Admittedly the modifications are a bit hacky due to my inexperience and the difference between Channels 1 and Channels 2 APIs, so maybe it would make more sense to fork the project into a Channels-1 version and a Channels-2 version to keep the code clean...

In addition to the changes in these files, the beatserver.py file requires additional lines to work correctly in Channels 2 (no changes needed if the user wants to keep working with Channels 1) which should be documented in the README file. This is an example beatserver.py file that works with Channels 2:

import os
from datetime import timedelta
from channels.layers import get_channel_layer

os.environ.setdefault("DJANGO_SETTINGS_MODULE", "PROJECT_NAME.settings")

BEAT_SCHEDULE = {
    'send-every-5-seconds': {
        'channel_name': 'CHANNEL_NAME',
        'schedule': timedelta(seconds=5),
        'message': {'type': 'type_of_message', 'message': 'message_text'}
    },
}

channel_layers = get_channel_layer()

and the server needs to be initialized with the following command:

beatserver PROJECT_NAME.beatconfig:application

Putting all this information in the beatconfig.py file is necessary to get the channel_layer. This is due to the fact that in Channels 2 if you try to import the PROJECT_NAME.asgi file as was done for Channels 1 you get a conflict between a the reactor imported in Server.py and an asyncioreactor used by Daphne. Daphne notices the conflict and proceeds to uninstall automatically the synchronous beatserver reactor which breaks beatserver, so I had to move the necessary information to get the channel_layers to the beatconfig.py file.... I know that this quite inelegant and you'll probably want to find a better solution before merging my changes, but I thought you'd prefer to at least have some "first working solution" you can improve upon :-).